+++ webapp/src/components/viewHeader/multiperson.scss
