+++ webapp/src/properties/multiperson/property.tsx
@@ -2,7 +2,7 @@
 // See LICENSE.txt for license information.
 import {IntlShape} from 'react-intl'
 
-import {PropertyType, PropertyTypeEnum} from '../types'
+import {PropertyType, PropertyTypeEnum, FilterValueType} from '../types'
 
 import MultiPerson from './multiperson'
 
@@ -11,4 +11,6 @@ export default class MultiPersonProperty extends PropertyType {
     name = 'MultiPerson'
     type = 'multiPerson' as PropertyTypeEnum
     displayName = (intl: IntlShape) => intl.formatMessage({id: 'PropertyType.MultiPerson', defaultMessage: 'Multi person'})
+    canFilter = true
+    filterValueType = 'person' as FilterValueType
 }
