+++ webapp/src/properties/person/property.tsx
@@ -2,7 +2,7 @@
 // See LICENSE.txt for license information.
 import {IntlShape} from 'react-intl'
 
-import {PropertyType, PropertyTypeEnum} from '../types'
+import {PropertyType, PropertyTypeEnum, FilterValueType} from '../types'
 
 import Person from './person'
 
@@ -11,4 +11,6 @@ export default class PersonProperty extends PropertyType {
     name = 'Person'
     type = 'person' as PropertyTypeEnum
     displayName = (intl: IntlShape) => intl.formatMessage({id: 'PropertyType.Person', defaultMessage: 'Person'})
+    canFilter = true
+    filterValueType = 'person' as FilterValueType
 }
