+++ webapp/src/properties/updatedBy/property.tsx
@@ -2,7 +2,7 @@
 // See LICENSE.txt for license information.
 import {IntlShape} from 'react-intl'
 
-import {PropertyType, PropertyTypeEnum} from '../types'
+import {PropertyType, PropertyTypeEnum, FilterValueType} from '../types'
 
 import UpdatedBy from './updatedBy'
 
@@ -12,4 +12,6 @@ export default class UpdatedByProperty extends PropertyType {
     type = 'updatedBy' as PropertyTypeEnum
     isReadOnly = true
     displayName = (intl: IntlShape) => intl.formatMessage({id: 'PropertyType.UpdatedBy', defaultMessage: 'Last updated by'})
+    canFilter = true
+    filterValueType = 'person' as FilterValueType
 }
